requirements.txt
streamlit==1.43.2
langchain==0.3.26
langchain-openai==0.3.28
python-dotenv==1.1.1
langchain-community